moving tests around to address files moved, skipping one of the tests
